requirements.txt
nemoguardrails==0.9
openai
langchain
langchain-openai
python-dotenv